Extracts maxCatsUpdateForm as a separate component
